requirements.txt
flask
flask-cors
flask-mysql
flask-sqlalchemy
flask-jwt-extended
python-dotenv
